+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 #### Features ####
 *   Add support for key modifiers for send_keys [Issue #420] (Sarah Mogin)
 *   Drag by offset support in native element (phoenixek12)
+*   Add access to properties of node's native element (Mike Souza)
 
 #### Bug fixes ####